% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishDiveR-package.R
\docType{package}
\name{FishDiveR-package}
\alias{FishDiveR}
\alias{FishDiveR-package}
\title{FishDiveR: Classify Aquatic Animal Behaviours from Vertical Movement Data}
\description{
Quantitatively analyse depth time-series data from pop-up satellite archival tags (PSATs) through the application of continuous wavelet transformation (CWT) combined with Principal Component Analysis (PCA), and k-means clustering. Import, crop, and plot depth time-depth records (TDRs). Using CWT to detect important signals within the non-stationary data, we create daily wavelet statistics to summarise vertical movements on different wavelet periods and combine with daily and diel depth statistics. Classify depth time-series with unsupervised k-means clustering into 24-hour periods of vertical movement behaviour with distinct patterns of vertical movement. Plot example days from each behaviour cluster, and plot the TDR coloured by cluster. Based on principals of combining CWT with k-means first developed by Sakamoto et al. (2009) \doi{10.1371/journal.pone.0005379} and redeveloped by Beale et al. (2024) \doi{ In Press}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/calvinsbeale/FishDiveR}
  \item Report bugs at \url{https://github.com/calvinsbeale/FishDiveR/issues}
}

}
\author{
\strong{Maintainer}: Calvin Beale \email{calvin.beale.8@gmail.com} (\href{https://orcid.org/0000-0002-3505-8929}{ORCID}) [copyright holder]

}
\keyword{internal}
