% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_k_clustering.R
\name{k_clustering}
\alias{k_clustering}
\title{Perform k-means}
\usage{
k_clustering(
  kmeans_data,
  standardise = TRUE,
  k,
  polygon = FALSE,
  output_folder = data_dir
)
}
\arguments{
\item{kmeans_data}{Data frame containing the combined PC scores and depth
statistics to perform k-means on. Output from the 'combine_data()'
function.}

\item{standardise}{TRUE or FALSE. Whether or not to standardise the data.
Defaults to TRUE.}

\item{k}{Numerical. Value of k to use for analysis.}

\item{polygon}{TRUE or FALSE. Plot polygons for cluster with more than 3 data
points. Defaults to FALSE.}

\item{output_folder}{Parent folder path with separate folders for each tag
data. E.g. "C:/Tag data". Defaults to 'data_dir'}
}
\value{
An object of class 'kmeans' containing the k-means clustering data
for the data frame. Additionally plots a 3D cluster plot of the top three
Principal Components.
}
\description{
\code{k_clustering} performs k-means clustering on the PC scores with the selected
value of k
}
\examples{
# Set file path
filepath <- system.file("extdata", package = "FishDiveR")

# Load kmeans_data
kmeans_data <- readRDS(file.path(filepath, "data/5_k-means/combined_stats.rds"))

# Run k_clustering function
kmeans_result <- k_clustering(
  kmeans_data = kmeans_data,
  standardise = TRUE,
  k = 4,
  polygon = FALSE,
  output_folder = tempdir()
)

}
